requirements-dev.txt
-r requirements.txt
django-test-without-migrations==0.6
ipdb==0.11
mixer==6.1.3
ipython==7.3.0
pandas==0.24.1
